allow decimal precision in draw_text y coordinate

+++ src/api/renderer.c
@@ -346,7 +346,7 @@ static int f_draw_text(lua_State *L) {
   size_t len;
   const char *text = luaL_checklstring(L, 2, &len);
   float x = luaL_checknumber(L, 3);
-  int y = luaL_checknumber(L, 4);
+  float y = luaL_checknumber(L, 4);
   RenColor color = checkcolor(L, 5, 255);
   x = rencache_draw_text(fonts, text, len, x, y, color);
   lua_pushnumber(L, x);
